Move SettingsPage logic to ViewModel
