+++ backlog/Views/SettingsPage.xaml
@@ -5,9 +5,10 @@
     xmlns:local="using:backlog.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
+    xmlns:muxc="using:Microsoft.UI.Xaml.Controls" 
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
 
     <Grid>
         <VisualStateManager.VisualStateGroups>
@@ -34,7 +35,7 @@
             </VisualStateGroup>
         </VisualStateManager.VisualStateGroups>
 
-        <ProgressRing x:Name="ProgRing" IsActive="{x:Bind ShowProgress}" />
+        <ProgressRing x:Name="ProgRing" IsActive="{Binding ShowProgress}" />
         <Pivot x:Name="mainPivot">
             <PivotItem Header="settings">
                 <ScrollViewer VerticalScrollBarVisibility="Auto">
@@ -63,11 +64,11 @@
                                     <x:String>Upcoming</x:String>
                                 </ComboBox>
                                 <muxc:RadioButtons x:Name="TileStyleRadioButtons" MaxColumns="2"
-                                                   Header="Live tile style" Margin="0,20,0,0" SelectedIndex="{x:Bind SelectedTileStyleIndex, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}">
+                                                   Header="Live tile style" Margin="0,20,0,0" SelectedIndex="{Binding SelectedTileStyleIndex, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}">
                                     <x:String>Peek Image</x:String>
                                     <x:String>Background Image</x:String>
                                 </muxc:RadioButtons>
-                                <Image x:Name="TileStyleImage" Source="{x:Bind TileStylePreviewImage, Mode=OneWay}" MaxHeight="100" 
+                                <Image x:Name="TileStyleImage" Source="{Binding TileStylePreviewImage, Mode=OneWay}" MaxHeight="100" 
                                        Stretch="Uniform" HorizontalAlignment="Left" />
                             </StackPanel>
 
@@ -75,14 +76,14 @@
                                 <TextBlock Text="Video playback" Margin="0,20,0,0" FontSize="18"/>
                                 <TextBlock Text="Settings for playing trailers for films, series and games." FontStyle="Italic"
                                            FontWeight="ExtraLight" TextWrapping="Wrap" />
-                                <ToggleSwitch x:Name="AutoplaySwitch" Header="Autoplay videos" IsOn="{x:Bind AutoplayVideos, Mode=TwoWay}" Margin="0,20,0,0" />
+                                <ToggleSwitch x:Name="AutoplaySwitch" Header="Autoplay videos" IsOn="{Binding AutoplayVideos, Mode=TwoWay}" Margin="0,20,0,0" />
                             </StackPanel>
 
                             <StackPanel x:Name="AccountPanel" Visibility="Collapsed" Margin="0,20,0,0">
                                 <TextBlock Text="Account" Margin="0,20,0,0" FontSize="18" />
                                 <TextBlock x:Name="UserNameText" Margin="0,10,0,20" FontStyle="Italic" />
                                 <muxc:PersonPicture x:Name="AccountPic" Margin="0,10,0,20" HorizontalAlignment="Left" />
-                                <Button x:Name="SignOutButton" Content="Sign out" Background="Red" Command="{x:Bind SignOut}" />
+                                <Button x:Name="SignOutButton" Content="Sign out" Background="Red" Command="{Binding SignOut}" />
                             </StackPanel>
 
                             <TextBlock x:Name="SignInText" Text="For the best experience, sign-in with your Microsoft Account. This will sync backlogs using OneDrive so you can access them on all of your Windows devices."
@@ -233,8 +234,8 @@
                             <TextBlock Text="If the app is crashing or you're unable to create Backlogs, or if anything odd is happening, sending me the logs will help me identify and fix the issue." 
                                    Margin="0,10,0,10" TextWrapping="Wrap" FontSize="15" />
                             <StackPanel Orientation="Horizontal">
-                                <Button x:Name="OpenLogsButton" Content="View logs" Command="{x:Bind OpenLogs}" Margin="0,10,0,10" />
-                                <Button x:Name="SendLogsButton" Content="Send logs" Command="{x:Bind SendLogs}" Margin="10,10,0,10" />
+                                <Button x:Name="OpenLogsButton" Content="View logs" Command="{Binding OpenLogs}" Margin="0,10,0,10" />
+                                <Button x:Name="SendLogsButton" Content="Send logs" Command="{Binding SendLogs}" Margin="10,10,0,10" />
                             </StackPanel>
                             <TextBlock Text="Have feedback? Want to report bugs or request a new feature? That's brilliant!" Margin="0,10,0,10" TextWrapping="Wrap" />
                             <TextBlock TextWrapping="Wrap" Margin="0,10,0,10" >
@@ -245,14 +246,14 @@
                             <TextBlock Text="OR"  Margin="20,20,0,20" />
                             <TextBlock Text="Send me an email" Margin="0,10,0,10" />
                             <ComboBox x:Name="IssueTypeComboBox" Margin="0,10,0,10" Header="Type" PlaceholderText="Pick a type"
-                                      SelectedValue="{x:Bind SelectedFeedbackType, Mode=TwoWay}">
+                                      SelectedValue="{Binding SelectedFeedbackType, Mode=TwoWay}">
                                 <x:String>Bug</x:String>
                                 <x:String>Feature Request</x:String>
                                 <x:String>Feedback</x:String>
                             </ComboBox>
                             <TextBox x:Name="MessageBox" PlaceholderText="Enter a description of your issue/feature" TextWrapping="Wrap"
-                                     MinHeight="80" Margin="0,10,0,10" Text="{x:Bind FeedbackText, Mode=TwoWay}" />
-                            <Button x:Name="SendButton" Content="Send" Command="{x:Bind SendFeedback}" />
+                                     MinHeight="80" Margin="0,10,0,10" Text="{Binding FeedbackText, Mode=TwoWay}" />
+                            <Button x:Name="SendButton" Content="Send" Command="{Binding SendFeedback}" />
                         </StackPanel>
 
                         <StackPanel x:Name="StoreLink" HorizontalAlignment="Right" Margin="0,20,20,0" >
