+++ backlog/Views/SettingsPage.xaml.cs
@@ -19,6 +19,7 @@ using MvvmHelpers.Commands;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.ServiceModel.Channels;
+using backlog.ViewModels;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
@@ -27,7 +28,7 @@ namespace backlog.Views
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
     /// </summary>
-    public sealed partial class SettingsPage : Page, INotifyPropertyChanged
+    public sealed partial class SettingsPage : Page
     {
         bool signedIn;
         static string MIT_LICENSE = "Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions: \n\nThe above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software. \n\nTHE SOFTWARE IS PROVIDED AS IS, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.";
@@ -40,97 +41,13 @@ namespace backlog.Views
         string ChangelogTitle = "New this version - 30 July, 2022";
         public string Version = Settings.Version;
 
-        private string selectedTheme = Settings.AppTheme;
-        private int selectedTileStyleIndex = Settings.TileStyle == "Peeking" ? 0 : 1;
-        private string tileStylePreviewImage = Settings.TileStyle == "Peeking" ? "ms-appx:///Assets/peeking-tile.png" :
-                "ms-appx:///Assets/background-tile.png";
-        private bool showProgress;
-        private string tileContent = Settings.TileContent;
 
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        public ICommand SendLogs { get; }
-        public ICommand OpenLogs { get; }
-        public ICommand SendFeedback { get; }
-        public ICommand SignOut { get;  }
-
-        public string SelectedTheme
-        {
-            get => selectedTheme;
-            set
-            {
-                selectedTheme = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(SelectedTheme)));
-                ChangeAppTheme();
-            }
-        }
-
-        public int SelectedTileStyleIndex
-        {
-            get => selectedTileStyleIndex;
-            set
-            {
-                selectedTileStyleIndex = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(SelectedTileStyleIndex)));
-                ChangeTileStyle();
-            }
-        }
-
-        public string SelectedTileContent
-        {
-            get => tileContent;
-            set
-            {
-                tileContent = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(SelectedTileContent)));
-                Settings.TileContent = value.ToString();
-            }
-        }
-
-        public string TileStylePreviewImage
-        {
-            get => tileStylePreviewImage;
-            set
-            {
-                tileStylePreviewImage = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(TileStylePreviewImage)));
-            }
-        }
-
-        public bool ShowProgress
-        {
-            get => showProgress;
-            set
-            {
-                showProgress = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(ShowProgress)));
-            }
-        }
-
-        public bool AutoplayVideos
-        {
-            get => Settings.AutoplayVideos;
-            set
-            {
-                Settings.AutoplayVideos = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(AutoplayVideos)));
-            }
-        }
-
-        public string SelectedFeedbackType { get; set; }
-
-        public string FeedbackText { get; set; }
-
+        public SettingsViewModel ViewModel { get; set; } = new SettingsViewModel();
 
         public SettingsPage()
         {
             this.InitializeComponent();
-            DataContext = this;
-
-            SendLogs = new AsyncCommand(SendLogsAsync);
-            OpenLogs = new AsyncCommand(ShowLogsAsync);
-            SendFeedback = new AsyncCommand(SendFeedbackAsync);
-            SignOut = new AsyncCommand(SignOutAsync);
+            this.DataContext = ViewModel;
 
             MyLicense.Text = GNU_LICENSE;
             WCTLicense.Text = MIT_LICENSE;
@@ -200,147 +117,6 @@ namespace backlog.Views
             }
 
             e.Handled = true;
-        }
-
-        /// <summary>
-        /// Change app theme on the fly and save it
-        /// </summary>
-        private void ChangeAppTheme()
-        {
-            var _selectedTheme = SelectedTheme;
-            if (_selectedTheme != null)
-            {
-                if (_selectedTheme == "System")
-                {
-                    _selectedTheme = "Default";
-                }
-                ThemeHelper.RootTheme = App.GetEnum<ElementTheme>(_selectedTheme);
-            }
-            Settings.AppTheme = SelectedTheme;
-        } 
-
-        /// <summary>
-        /// Opens email client to send logs
-        /// </summary>
-        /// <returns></returns>
-        private async Task SendLogsAsync()
-        {
-            ShowProgress = true;
-            EmailMessage emailMessage = new EmailMessage();
-            emailMessage.To.Add(new EmailRecipient("surya.sk05@outlook.com"));
-            emailMessage.Subject = "Logs from Backlogs";
-            StringBuilder body = new StringBuilder();
-            body.AppendLine("*Enter a brief description of your issue here*");
-            body.AppendLine("\n\n\n");
-            body.AppendLine("Logs:");
-            var logList = await Logger.GetLogsAsync();
-            foreach (var log in logList)
-            {
-                body.AppendLine(log.ToString());
-            }
-            emailMessage.Body = body.ToString();
-            await EmailManager.ShowComposeNewEmailAsync(emailMessage);
-            ShowProgress = false;
-        }
-
-        /// <summary>
-        /// Opens a content dialog that shows logs
-        /// </summary>
-        /// <returns></returns>
-        private async Task ShowLogsAsync()
-        {
-            var logs = await Logger.GetLogsAsync();
-            ContentDialog contentDialog = new ContentDialog()
-            {
-                Title = "Logs",
-                Content = new ListView()
-                {
-                    ItemsSource = logs,
-                    HorizontalContentAlignment = HorizontalAlignment.Stretch,
-                    IsItemClickEnabled = false,
-                    SelectionMode = ListViewSelectionMode.None
-                },
-                CloseButtonText = "Close"
-            };
-            await contentDialog.ShowAsync();
-        }
-
-        /// <summary>
-        /// Send user typed feedback
-        /// </summary>
-        /// <returns></returns>
-        private async Task SendFeedbackAsync()
-        {
-            if(string.IsNullOrEmpty(SelectedFeedbackType) || string.IsNullOrEmpty(FeedbackText))
-            {
-                await ShowFeedbackInputErrorAsync();
-            }
-            else
-            {
-                await SendFeedbackEmailAsync();
-            }
-        }
-
-        /// <summary>
-        /// Show error dialog
-        /// </summary>
-        /// <returns></returns>
-        private async Task ShowFeedbackInputErrorAsync()
-        {
-            ContentDialog contentDialog = new ContentDialog
-            {
-                Title = "Insufficient data",
-                Content = "Please fill in both the fields",
-                CloseButtonText = "Ok"
-            };
-            ContentDialogResult result = await contentDialog.ShowAsync();
-        }
-
-        /// <summary>
-        /// Open the user's default email client
-        /// </summary>
-        /// <returns></returns>
-        private async Task SendFeedbackEmailAsync()
-        {
-            ShowProgress = true;
-            EmailMessage emailMessage = new EmailMessage();
-            emailMessage.Subject = "[Backlogs] " + IssueTypeComboBox.SelectedItem.ToString();
-            emailMessage.Body = MessageBox.Text;
-            emailMessage.To.Add(new EmailRecipient("surya.sk05@outlook.com"));
-            await EmailManager.ShowComposeNewEmailAsync(emailMessage);
-            ShowProgress = false;
-        }
-
-        /// <summary>
-        /// Sign the user out of MSAL
-        /// </summary>
-        /// <returns></returns>
-        private async Task SignOutAsync()
-        {
-            ContentDialog contentDialog = new ContentDialog
-            {
-                Title = "Sign out?",
-                Content = "You will no longer have access to your backlogs, and new ones will no longer be synced",
-                PrimaryButtonText = "Yes",
-                CloseButtonText = "No"
-            };
-            ContentDialogResult result = await contentDialog.ShowAsync();
-            if (result == ContentDialogResult.Primary)
-            {
-                await MSAL.SignOut();
-                Settings.IsSignedIn = false;
-                Frame.Navigate(typeof(MainPage));
-            }
-        }
-
-        /// <summary>
-        /// Change tile style
-        /// </summary>
-        private void ChangeTileStyle()
-        {
-            TileStylePreviewImage = selectedTileStyleIndex == 0 ? "ms-appx:///Assets/peeking-tile.png" :
-    "ms-appx:///Assets/background-tile.png";
-            Settings.TileStyle = selectedTileStyleIndex == 0 ? "Peeking" : "Background";
         }
     }
 }
